feat: add RunCountTest to run non-determinism checks
